refactor(ci): split the /e2e workflow into modules, parameterise server storage
Addresses review feedback on #5586.

1. `pullRequestsCommandE2e.wac.ts` had grown to 479 lines holding constants,
   shared step groups and two job factories. Split into `wac/e2e/`:

     constants.ts          directories, ports, URLs
     sharedSteps.ts        the step groups every job in the workflow reuses
     createCypressJobs.ts  the AWS-deployed variants
     createServerJobs.ts   the self-hosted variants
     index.ts              re-exports

   The `.wac.ts` file is now 88 lines of workflow assembly. Helpers deliberately
   do NOT end in `.wac.ts`, so ghawac does not treat them as workflows.

   `createCheckoutPrSteps` was duplicated verbatim in the /e2e and /vitest
   sources; it now lives in `steps/` and takes a working directory. Both
   workflows use it and their generated YAML is byte-identical to before.

2. `createServerJobs` now takes a storage backend rather than hardcoding SQLite:

     createServerJobs("sqlite")    // registered
     createServerJobs("postgres")  // ready, not registered

   The postgres path adds a `postgres:17` service container with a pg_isready
   health check (without it the job can reach `webiny build` before Postgres
   accepts connections) and the WEBINY_PG_* env matching the postgres template's
   .env.example defaults. Verified it renders correctly, then left unregistered:
   the server hosting type is still ALPHA, so prove one variant green before
   doubling the surface. Enabling it is one line.

3. `scripts/setupCypress.js` was accumulating flags and branching inline. It now
   has one value factory per hosting type - `aws` reads the deployed stacks'
   output, `server` takes the URLs directly - selected by `--hosting-type`.
   Adding a hosting type means adding a factory, not threading another flag
   through the substitution code.

   Two bugs fixed while restructuring: the unknown-hosting-type message passed
   `Array.map`'s index straight into chalk ("expected aws 0 aws,server"), and a
   missing-argument exit could leave a half-substituted cypress.config.ts on
   disk. Values are now resolved before the file is touched.

No pre-existing job changed: every job in the generated YAML is byte-identical
to `next` except the added `e2e-server-sqlite`.
