+++ .github/workflows/wac/pullRequestsCommandVitest.wac.ts
@@ -1,5 +1,6 @@
 import { NormalJob } from "github-actions-wac";
 import {
+    createCheckoutPrSteps,
     createGlobalBuildCacheSteps,
     createInstallBuildSteps,
     createRunBuildArtifactDownloadSteps,
@@ -40,21 +41,6 @@ const runBuildCacheUploadSteps = createRunBuildArtifactUploadSteps({
 const runBuildCacheDownloadSteps = createRunBuildArtifactDownloadSteps({
     workingDirectory: DIR_WEBINY_JS
 });
-
-const createCheckoutPrSteps = () =>
-    [
-        {
-            name: "Checkout Pull Request",
-            "working-directory": DIR_WEBINY_JS,
-            // Detach onto the SHA `baseBranch` resolved, so every job in this run builds and
-            // tests the same commit even if the PR is pushed to mid-run.
-            run: [
-                "gh pr checkout ${{ github.event.issue.number }}",
-                "git checkout --detach ${{ needs.baseBranch.outputs.pr-sha }}"
-            ].join("\n"),
-            env: { GITHUB_TOKEN: "${{ secrets.GH_TOKEN }}" }
-        }
-    ] as NonNullable<NormalJob["steps"]>;
 
 // Live status table shown in the PR comment. Rows are updated in place as each
 // group progresses (Queued -> Running -> Passed/Failed) by the per-group jobs,
@@ -166,7 +152,7 @@ const createVitestTestsJobs = (storageOps?: AbstractStorageOps) => {
                 // existing one, or changed a package's testing/sharding config would have those
                 // changes silently ignored - while a PR that deleted a package would still get a
                 // test job for it.
-                ...createCheckoutPrSteps(),
+                ...createCheckoutPrSteps({ workingDirectory: DIR_WEBINY_JS }),
                 {
                     id: "list-vitest-test-commands",
                     name: "List Vitest Test Commands",
@@ -202,7 +188,7 @@ const createVitestTestsJobs = (storageOps?: AbstractStorageOps) => {
             awsAuth: storageOps && storageOps.id === "ddb-os,ddb",
             checkout: { path: DIR_WEBINY_JS },
             steps: [
-                ...createCheckoutPrSteps(),
+                ...createCheckoutPrSteps({ workingDirectory: DIR_WEBINY_JS }),
                 ...yarnCacheSteps,
                 ...runBuildCacheDownloadSteps,
                 ...installBuildSteps,
@@ -274,7 +260,7 @@ export const pullRequestsCommandVitest = createSlashCommandWorkflow({
             checkout: { path: DIR_WEBINY_JS },
             "runs-on": BUILD_PACKAGES_RUNNER,
             steps: [
-                ...createCheckoutPrSteps(),
+                ...createCheckoutPrSteps({ workingDirectory: DIR_WEBINY_JS }),
                 ...yarnCacheSteps,
                 ...globalBuildCacheSteps,
                 ...installBuildSteps,
