+++ .github/workflows/wac/steps/index.ts
@@ -1,3 +1,4 @@
+export * from "./createCheckoutPrSteps.js";
 export * from "./createDeployWebinySteps.js";
 export * from "./createSetupVerdaccioSteps.js";
 export * from "./createInstallBuildSteps.js";
